Store 생성 및 누구나 전체 Store 조회 구현
