+++ src/main/java/com/delivery/domain/store/controller/AdminStoreController.java
@@ -1,0 +1,4 @@
+package com.delivery.domain.store.controller;
+
+public class AdminStoreController {
+}
