+++ src/main/java/com/delivery/domain/store/dto/StoreCreateRequestDto.java
@@ -1,5 +1,6 @@
-package com.delivery.store.dto;
+package com.delivery.domain.store.dto;
 
+import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
@@ -7,7 +8,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class StoreUpdateRequestDto {
+public class StoreCreateRequestDto {
 
     @NotBlank(message = "가게명은 필수입니다.")
     private String name;
@@ -27,4 +28,6 @@ public class StoreUpdateRequestDto {
     @NotNull(message = "카테고리는 필수입니다.")
     private Integer categoryId;
 
+    @NotNull(message = "점주 ID는 필수입니다.")
+    private Long ownerUserId;
 }
