+++ src/main/java/com/delivery/domain/store/repository/StoreCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.delivery.store.repository;
+package com.delivery.domain.store.repository;
 
-import com.delivery.store.entity.StoreCategory;
+import com.delivery.domain.store.entity.StoreCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreCategoryRepository extends JpaRepository<StoreCategory,Integer> {
