+++ src/main/java/com/delivery/domain/store/service/StoreServiceImpl.java
@@ -1,13 +1,13 @@
-package com.delivery.store.service;
+package com.delivery.domain.store.service;
 
-import com.delivery.store.dto.StoreCreateRequestDto;
-import com.delivery.store.dto.StoreResponseDto;
-import com.delivery.store.entity.Store;
-import com.delivery.store.entity.StoreCategory;
-import com.delivery.store.repository.StoreCategoryRepository;
-import com.delivery.store.repository.StoreRepository;
-import com.delivery.user.entity.User;
-import com.delivery.user.entity.UserRoleEnum;
+import com.delivery.domain.store.dto.StoreCreateRequestDto;
+import com.delivery.domain.store.dto.StoreResponseDto;
+import com.delivery.domain.store.entity.Store;
+import com.delivery.domain.store.entity.StoreCategory;
+import com.delivery.domain.store.repository.StoreCategoryRepository;
+import com.delivery.domain.store.repository.StoreRepository;
+import com.delivery.domain.user.entity.User;
+import com.delivery.domain.user.entity.UserRoleEnum;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -16,11 +16,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.Locale;
-
 @Service
 @RequiredArgsConstructor
-public class storeServiceImpl implements storeService {
+public class StoreServiceImpl implements StoreService {
 
     private final StoreRepository storeRepository;
     private final StoreCategoryRepository storeCategoryRepository;
@@ -46,18 +44,29 @@ public class storeServiceImpl implements storeService {
                 user.getUserId()
         );
         storeRepository.save(store);
-        return new StroreResponseDto(store);
+        return new StoreResponseDto(store);
     }
 
     // 2. OWNER, MASTER - 가게 수정
 
     // 3. OWNER, MASTER - 가게 삭제
 
-    // 4. 누구나 전체 가게 목록 조회
+    // 4. OWNER, MASTER - 본인 가게 조회
 
-    // 5. 누구나 가게 상세 조회
+    // 5. 누구나 전체 가게 목록 조회
+    @Override
+    public Page<StoreResponseDto> getStores(int page, int size, String sortBy, boolean isAsc) {
+        Sort.Direction dir = isAsc ? Sort.Direction.ASC : Sort.Direction.DESC;
+        Pageable pageable = PageRequest.of(page, size, Sort.by(dir, sortBy));
 
-    // 6. OWNER, MASTER - 본인 가게 조회
+        Page<Store> stores = storeRepository.findAll(pageable);
+        return stores.map(StoreResponseDto::new);
+    }
+
+    // 6. 누구나 가게 상세 조회
+
 
 
 }
+
+
