+++ src/main/java/com/delivery/domain/test/TestEntity.java
@@ -14,7 +14,7 @@ import lombok.*;
 class TestEntity {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue
     private Long id;
 
     private String value;
