--- src/main/java/com/delivery/store/controller/AdminStoreController.java
@@ -1,4 +1,0 @@
-package com.delivery.store.controller;
-
-public class AdminStoreController {
-}
