--- src/main/java/com/delivery/store/repository/StoreRepository.java
@@ -1,14 +1,0 @@
-package com.delivery.store.repository;
-
-import com.delivery.store.entity.Store;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface StoreRepository extends JpaRepository<Store,UUID> {
-
-}
